refactor: rename files in tag, build-runtime-services, tree-api (#1201)
- tag: ports.ts → types.ts
- build-runtime-services: eventEmission.ts → eventEmissionUtils.ts
- tree-api: command-payloads.ts → commandPayloadTypes.ts,
  dialog-state.ts → dialogStateTypes.ts,
  undo-state-events.ts → undoStateEventTypes.ts

Rollback: git revert <hash> --no-edit
