+++ packages/tag/src/index.ts
@@ -1,4 +1,4 @@
 export * from './TagService.js';
-export * from './ports.js';
+export * from './types.js';
 export * from './capability.js';
